logging(validation): log excluded savePaths from linking test (#850)

+++ src/clients/TorrentClient.ts
@@ -16,7 +16,7 @@ import {
 } from "../constants.js";
 import { getRuntimeConfig } from "../runtimeConfig.js";
 import { Searchee, SearcheeWithInfoHash } from "../searchee.js";
-import { wait } from "../utils.js";
+import { formatAsList, wait } from "../utils.js";
 import Deluge from "./Deluge.js";
 import QBittorrent from "./QBittorrent.js";
 import RTorrent from "./RTorrent.js";
@@ -97,25 +97,41 @@ export async function validateSavePaths(
 	const { blockList, linkDir } = getRuntimeConfig();
 	logger.info(`Validating all existing torrent save paths...`);
 
-	const entry = searchees.find((s) => !infoHashPathMap.has(s.infoHash));
-	if (entry) {
+	const entryDir = searchees.find((s) => !infoHashPathMap.has(s.infoHash));
+	if (entryDir) {
 		logger.warn(
-			`Not all torrents from torrentDir are in the torrent client (missing ${entry.name} [${entry.infoHash}]): https://www.cross-seed.org/docs/basics/options#torrentdir`,
+			`Not all torrents from torrentDir are in the torrent client (missing ${entryDir.name} [${entryDir.infoHash}]): https://www.cross-seed.org/docs/basics/options#torrentdir`,
 		);
 	}
-	if (infoHashPathMap.size !== searchees.length) {
+	const searcheeInfoHashes = new Set(searchees.map((s) => s.infoHash));
+	const entryClient = Array.from(infoHashPathMap.keys()).find(
+		(infoHash) => !searcheeInfoHashes.has(infoHash),
+	);
+	if (entryClient) {
 		logger.warn(
-			"Could not ensure all torrents from the torrent client are in torrentDir (likely incorrect torrentDir or torrents with infohash_v2): https://www.cross-seed.org/docs/basics/options#torrentdir",
+			`Could not ensure all torrents from the torrent client are in torrentDir (missing ${entryClient} with savePath ${infoHashPathMap.get(entryClient)}): https://www.cross-seed.org/docs/basics/options#torrentdir`,
 		);
 	}
 	if (!linkDir) return;
 
+	const removedSavePaths = new Set<string>();
 	for (const searchee of searchees) {
 		if (findBlockedStringInReleaseMaybe(searchee, blockList)) {
-			infoHashPathMap.delete(searchee.infoHash);
+			if (infoHashPathMap.has(searchee.infoHash)) {
+				removedSavePaths.add(infoHashPathMap.get(searchee.infoHash)!);
+				infoHashPathMap.delete(searchee.infoHash);
+			}
 		}
 	}
-	for (const savePath of new Set(infoHashPathMap.values())) {
+	const uniqueSavePaths = new Set(infoHashPathMap.values());
+	const ignoredSavePaths = Array.from(removedSavePaths).filter(
+		(savePath) => !uniqueSavePaths.has(savePath),
+	);
+	logger.verbose(
+		`Excluded save paths from linking test due to blockList: ${formatAsList(ignoredSavePaths, { sort: true, type: "unit" })}`,
+	);
+
+	for (const savePath of uniqueSavePaths) {
 		if (ABS_WIN_PATH_REGEX.test(savePath) === (path.sep === "/")) {
 			throw new CrossSeedError(
 				`Cannot use linkDir with cross platform cross-seed and torrent client: ${savePath}`,
